requirements.txt
datasets
accelerate
openai
sentencepiece
python-dotenv
protobuf==3.20.0
evaluate
wandb
torch
pandas
numpy
scikit-learn
utils
git+https://github.com/huggingface/transformers.git
huggingface_hub
tenacity
rq
redis
loralib
bitsandbytes
git+https://github.com/huggingface/peft.git
gradio
appdirs
